renamed RapSheetAnalyzer back to analyzer so it is more consistent with project style.
