+++ src/backend/expungeservice/__init__.py
@@ -1,4 +1,4 @@
 from . import app
 from .app import create_app
 from . import endpoints
-from . import RapSheetAnalyzer
+from . import analyzer
